feat: split testing surface harness subpaths (#573)
## Context

Implements [TRL-757](https://linear.app/outfitter/issue/TRL-757/split-ontrailstesting-surface-harnesses-behind-subpaths): isolate `@ontrails/testing`'s root from optional surface peers by moving CLI/MCP/HTTP harnesses and the surface-parity helper behind explicit subpath exports. Driven by the PatchOS dogfood finding that the root testing barrel pulled `@ontrails/http` source into downstream `tsc`.

Closes [TRL-757](https://linear.app/outfitter/issue/TRL-757/split-ontrailstesting-surface-harnesses-behind-subpaths).

## Changes

- `packages/testing` `exports` map: root contract helpers stay at `.`; surface harnesses move to `./cli`, `./mcp`, `./http`, `./surface-parity`; `testAllEstablished` moves to `./established`.
- `peerDependencies` keeps `@ontrails/cli`, `@ontrails/http`, `@ontrails/mcp`; `peerDependenciesMeta.optional: true` marks them optional so apps that only use contract helpers do not need to install the surface peers.
- Localized surface-only types out of root `src/types.ts` and root `src/all.ts`; `src/all-established.ts` is the new owner of `testAllEstablished`.
- New regression: `packages/testing/src/__tests__/public-subpaths.test.ts` uses `--traceResolution` to prove the root boundary and typechecks the subpath surface.
- Consumer/scaffolder migration:
  - `apps/trails-demo/__tests__/examples.test.ts` and `surface-parity.test.ts` switched to `@ontrails/testing/established` and `@ontrails/testing/surface-parity`.
  - `apps/trails/src/trails/add-verify.ts` scaffold output now emits `from '@ontrails/testing/established'`.
- `knip.config.ts` adds a `packages/testing` workspace override with `ignoreDependencies: ['@ontrails/cli', '@ontrails/http', '@ontrails/mcp']`. Comment in the config justifies the carve-out: the subpath layout intentionally references the optional peers; knip's "Referenced optional peerDependencies" hint is expected here.
- Docs/plugin guidance updated: `docs/testing.md`, `docs/api-reference.md`, `packages/testing/README.md`, `plugin/skills/trails/SKILL.md`, `plugin/skills/trails/references/testing-patterns.md`, `plugin/skills/trails/references/http-surface.md`, `plugin/skills/trails/references/architecture.md`.
- `.changeset/testing-surface-subpaths.md` (`minor` for `@ontrails/testing`).

## Verification

- `bun run --cwd packages/testing typecheck` → exit 0.
- Targeted six-file `bun test`: `public-subpaths.test.ts`, `all.test.ts`, `harness-cli.test.ts`, `harness-http.test.ts`, `harness-mcp.test.ts`, `surface-parity.test.ts` → 37 pass, 0 fail.
- `bun run docs:snippets` → 21 README files typecheck (`packages/testing/README.md` included).
- `bun run docs:api-examples` → minimum public API example coverage passes.
- `bun run warden:skills:check` → exit 0.
- `bunx markdownlint-cli2` on changed docs/plugin/changeset files → 0 errors.
- `bun run docs:links` → 119 files.
- `bun run publish:check` → all publishable workspaces pack-check pass; `@ontrails/testing` tarball includes the new subpath source files.
- `bun run format:check` → 0 warnings/errors.
- `bun run dead-code` → clean after the `packages/testing` knip override.
- `bun run test` (full repo) → 37 packages all pass; `trails-demo` 74 pass after the consumer-import migration.
- `git diff --check` → clean.

## Risks / Rollout

User-visible API change for consumers of `@ontrails/testing`. Migration is a one-line subpath swap per import. Migration guidance lives in the TRL-760 beta.15 → beta.18 doc (top of stack) and in the package README. The optional `peerDependenciesMeta` does not change install behavior for apps that already had the peers installed; apps that only use root contract helpers can drop the surface peers from their package.json.

## Follow-ups

None on this branch. The consumer-side migration story is captured in the TRL-760 migration guide.
