+++ packages/testing/package.json
@@ -12,6 +12,11 @@
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
+    "./cli": "./src/cli.ts",
+    "./established": "./src/all-established.ts",
+    "./http": "./src/http.ts",
+    "./mcp": "./src/mcp.ts",
+    "./surface-parity": "./src/surface-parity.ts",
     "./package.json": "./package.json"
   },
   "scripts": {
@@ -32,5 +37,16 @@
     "@ontrails/mcp": "workspace:^",
     "@ontrails/observe": "workspace:^",
     "zod": "catalog:"
+  },
+  "peerDependenciesMeta": {
+    "@ontrails/cli": {
+      "optional": true
+    },
+    "@ontrails/http": {
+      "optional": true
+    },
+    "@ontrails/mcp": {
+      "optional": true
+    }
   }
 }
