+++ packages/testing/src/all.ts
@@ -8,16 +8,12 @@
 import { describe, expect, test } from 'bun:test';
 
 import type { Topo, TrailContext } from '@ontrails/core';
-import { validateEstablishedTopo, validateTopo } from '@ontrails/core';
+import { validateTopo } from '@ontrails/core';
 
-import { createCliHarness } from './harness-cli.js';
-import { createHttpHarness } from './harness-http.js';
-import { createMcpHarness } from './harness-mcp.js';
 import { testContracts } from './contracts.js';
 import type { TestExecutionOptions } from './context.js';
 import { testDetours } from './detours.js';
 import { testExamples } from './examples.js';
-import type { TestAllEstablishedOptions } from './types.js';
 
 /**
  * Run the full contract test suite for a Topo.
@@ -40,7 +36,7 @@ import type { TestAllEstablishedOptions } from './types.js';
  * testAll(graph);
  * ```
  */
-type TestAllInput =
+export type TestAllInput =
   | Partial<TrailContext>
   | TestExecutionOptions
   | (() => Partial<TrailContext> | TestExecutionOptions);
@@ -74,7 +70,7 @@ const assertValidTopo = (result: ReturnType<typeof validateTopo>): void => {
   }
 };
 
-const registerContractSuite = (
+export const registerContractSuite = (
   topo: Topo,
   ctxOrFactory: TestAllInput | undefined,
   validate: (topo: Topo) => ReturnType<typeof validateTopo>
@@ -95,127 +91,4 @@ const registerContractSuite = (
 
 export const testAll = (topo: Topo, ctxOrFactory?: TestAllInput): void => {
   registerContractSuite(topo, ctxOrFactory, validateTopo);
-};
-
-type EstablishedInput =
-  | Partial<TrailContext>
-  | TestAllEstablishedOptions
-  | (() => Partial<TrailContext> | TestAllEstablishedOptions);
-
-const isEstablishedOptions = (
-  input: Partial<TrailContext> | TestAllEstablishedOptions | undefined
-): input is TestAllEstablishedOptions =>
-  input !== undefined &&
-  (Object.hasOwn(input, 'cli') ||
-    Object.hasOwn(input, 'createPermit') ||
-    Object.hasOwn(input, 'ctx') ||
-    Object.hasOwn(input, 'http') ||
-    Object.hasOwn(input, 'mcp') ||
-    Object.hasOwn(input, 'resources') ||
-    Object.hasOwn(input, 'strictPermits'));
-
-const normalizeEstablishedOptions = (
-  input?: Partial<TrailContext> | TestAllEstablishedOptions
-): TestAllEstablishedOptions =>
-  isEstablishedOptions(input) ? input : { ctx: input };
-
-const toExecutionOptions = (
-  options: TestAllEstablishedOptions
-): TestExecutionOptions => ({
-  ...(options.createPermit === undefined
-    ? {}
-    : { createPermit: options.createPermit }),
-  ...(options.ctx === undefined ? {} : { ctx: options.ctx }),
-  ...(options.resources === undefined ? {} : { resources: options.resources }),
-  ...(options.strictPermits === undefined
-    ? {}
-    : { strictPermits: options.strictPermits }),
-});
-
-const toCliHarnessOptions = (
-  topo: Topo,
-  options: TestAllEstablishedOptions
-) => {
-  const cliOptions = {
-    graph: topo,
-    ...options.cli,
-  };
-
-  if (options.ctx !== undefined) {
-    cliOptions.ctx = options.ctx;
-  }
-
-  return cliOptions;
-};
-
-const toMcpHarnessOptions = (
-  topo: Topo,
-  options: TestAllEstablishedOptions
-) => ({
-  graph: topo,
-  ...options.mcp,
-});
-
-const toHttpHarnessOptions = (
-  topo: Topo,
-  options: TestAllEstablishedOptions
-) => {
-  const httpOptions = {
-    graph: topo,
-    ...options.http,
-  };
-
-  if (options.ctx !== undefined) {
-    httpOptions.ctx = options.ctx;
-  }
-
-  return httpOptions;
-};
-
-const registerEstablishedSurfaceSuite = (
-  topo: Topo,
-  resolveInput: () =>
-    | Partial<TrailContext>
-    | TestAllEstablishedOptions
-    | undefined
-): void => {
-  describe('surfaces', () => {
-    test('CLI projection validates established topo', () => {
-      const options = normalizeEstablishedOptions(resolveInput());
-      expect(() =>
-        createCliHarness(toCliHarnessOptions(topo, options))
-      ).not.toThrow();
-    });
-
-    test('MCP projection validates established topo', () => {
-      const options = normalizeEstablishedOptions(resolveInput());
-      expect(() =>
-        createMcpHarness(toMcpHarnessOptions(topo, options))
-      ).not.toThrow();
-    });
-
-    test('HTTP projection validates established topo', () => {
-      const options = normalizeEstablishedOptions(resolveInput());
-      expect(() =>
-        createHttpHarness(toHttpHarnessOptions(topo, options))
-      ).not.toThrow();
-    });
-  });
-};
-
-export const testAllEstablished = (
-  topo: Topo,
-  optionsOrFactory?: EstablishedInput
-): void => {
-  const resolveInput =
-    typeof optionsOrFactory === 'function'
-      ? optionsOrFactory
-      : () => optionsOrFactory;
-
-  registerContractSuite(
-    topo,
-    () => toExecutionOptions(normalizeEstablishedOptions(resolveInput())),
-    validateEstablishedTopo
-  );
-  registerEstablishedSurfaceSuite(topo, resolveInput);
 };
