+++ packages/testing/src/cli.ts
@@ -1,0 +1,6 @@
+export { createCliHarness } from './harness-cli.js';
+export type {
+  CliHarness,
+  CliHarnessOptions,
+  CliHarnessResult,
+} from './harness-cli.js';
