+++ packages/testing/src/harness-http.ts
@@ -6,18 +6,96 @@
  */
 
 import { deriveHttpRoutes } from '@ontrails/http';
-import type { HttpMethod, HttpRouteDefinition } from '@ontrails/http';
-import type { TrailContext, TrailContextInit } from '@ontrails/core';
+import type {
+  DeriveHttpRoutesOptions,
+  HttpHeaderSource,
+  HttpMethod,
+  HttpRouteDefinition,
+} from '@ontrails/http';
+import type { Topo, TrailContext, TrailContextInit } from '@ontrails/core';
 import { NotFoundError, projectPublicSurfaceError } from '@ontrails/core';
 
 import { mergeTestContext } from './context.js';
-import type {
-  HttpHarness,
-  HttpHarnessOptions,
+
+/** Options for creating an HTTP harness. */
+export interface HttpHarnessOptions extends DeriveHttpRoutesOptions {
+  readonly ctx?: Partial<TrailContext> | undefined;
+  readonly graph: Topo;
+}
+
+export interface HttpHarnessRequest {
+  readonly abortSignal?: AbortSignal | undefined;
+  readonly body?: unknown | undefined;
+  readonly headers?: HttpHeaderSource | undefined;
+  readonly method: HttpMethod;
+  readonly path: string;
+  readonly query?: Record<string, unknown> | undefined;
+  readonly requestId?: string | undefined;
+}
+
+export interface HttpHarnessRequestOptions extends Omit<
   HttpHarnessRequest,
-  HttpHarnessRequestOptions,
-  HttpHarnessResult,
-} from './types.js';
+  'body' | 'method' | 'path' | 'query'
+> {
+  readonly query?: Record<string, unknown> | undefined;
+}
+
+/** A test harness for HTTP route projections. */
+export interface HttpHarness {
+  /** Execute a raw HTTP-style harness request. */
+  request(request: HttpHarnessRequest): Promise<HttpHarnessResult>;
+  /** Execute a GET request, reading input from query params. */
+  get(
+    path: string,
+    query?: Record<string, unknown>,
+    options?: HttpHarnessRequestOptions
+  ): Promise<HttpHarnessResult>;
+  /** Execute a POST request, reading input from the JSON-like body value. */
+  post(
+    path: string,
+    body?: unknown,
+    options?: HttpHarnessRequestOptions
+  ): Promise<HttpHarnessResult>;
+  /** Execute a PUT request. */
+  put(
+    path: string,
+    body?: unknown,
+    options?: HttpHarnessRequestOptions
+  ): Promise<HttpHarnessResult>;
+  /** Execute a PATCH request. */
+  patch(
+    path: string,
+    body?: unknown,
+    options?: HttpHarnessRequestOptions
+  ): Promise<HttpHarnessResult>;
+  /** Execute a DELETE request. */
+  delete(
+    path: string,
+    body?: unknown,
+    options?: HttpHarnessRequestOptions
+  ): Promise<HttpHarnessResult>;
+}
+
+export interface HttpHarnessErrorBody {
+  readonly error: {
+    readonly category: string;
+    readonly code: string;
+    readonly message: string;
+  };
+}
+
+export interface HttpHarnessSuccessBody {
+  readonly data: unknown;
+}
+
+/** The result of an HTTP harness request. */
+export interface HttpHarnessResult {
+  readonly body: HttpHarnessErrorBody | HttpHarnessSuccessBody;
+  readonly data?: unknown | undefined;
+  readonly error?: HttpHarnessErrorBody['error'] | undefined;
+  readonly ok: boolean;
+  readonly status: number;
+}
 
 const TEST_ORIGIN = 'http://ontrails.test';
 
@@ -163,7 +241,7 @@ const executeRoute = async (
  *
  * @example
  * ```ts
- * import { createHttpHarness } from '@ontrails/testing';
+ * import { createHttpHarness } from '@ontrails/testing/http';
  *
  * const http = createHttpHarness({ graph });
  * const result = await http.get('/entity/show', { name: 'Alpha' });
