+++ packages/testing/src/http.ts
@@ -1,0 +1,10 @@
+export { createHttpHarness } from './harness-http.js';
+export type {
+  HttpHarness,
+  HttpHarnessErrorBody,
+  HttpHarnessOptions,
+  HttpHarnessRequest,
+  HttpHarnessRequestOptions,
+  HttpHarnessResult,
+  HttpHarnessSuccessBody,
+} from './harness-http.js';
