+++ packages/testing/src/index.ts
@@ -1,15 +1,10 @@
 // Contract-driven testing
 export { testAll } from './all.js';
-export { testAllEstablished } from './all.js';
 export { testExamples } from './examples.js';
 export { testCrosses } from './crosses.js';
 export { testTrail } from './trail.js';
 export { testContracts } from './contracts.js';
 export { testDetours } from './detours.js';
-export {
-  runSurfaceParityExample,
-  testSurfaceParity,
-} from './surface-parity.js';
 
 // Assertions
 export {
@@ -34,11 +29,6 @@ export {
 } from './context.js';
 export { createTestLogger } from './logger.js';
 
-// Surface harnesses
-export { createCliHarness } from './harness-cli.js';
-export { createHttpHarness } from './harness-http.js';
-export { createMcpHarness } from './harness-mcp.js';
-
 // Types
 export type { CreateCrossContextOptions } from './context.js';
 export type {
@@ -52,26 +42,7 @@ export type {
   CrossScenario,
   RefToken,
   ScenarioStep,
-  TestAllEstablishedOptions,
   TestScenario,
   TestLogger,
   TestTrailContextOptions,
-  CliHarness,
-  CliHarnessOptions,
-  CliHarnessResult,
-  HttpHarness,
-  HttpHarnessErrorBody,
-  HttpHarnessOptions,
-  HttpHarnessRequest,
-  HttpHarnessRequestOptions,
-  HttpHarnessResult,
-  HttpHarnessSuccessBody,
-  McpHarness,
-  McpHarnessOptions,
-  McpHarnessResult,
-  NormalizedSurfaceParityResult,
-  SurfaceParityExclusion,
-  SurfaceParityOptions,
-  SurfaceParitySurface,
 } from './types.js';
-export type { SurfaceParityComparison } from './surface-parity.js';
