+++ packages/testing/src/mcp.ts
@@ -1,0 +1,6 @@
+export { createMcpHarness } from './harness-mcp.js';
+export type {
+  McpHarness,
+  McpHarnessOptions,
+  McpHarnessResult,
+} from './harness-mcp.js';
